ISSUE #1177 - Handle disable state for tabs and refactor code
