+++ frontend/modules/teamspace/teamspace.helpers.ts
@@ -1,0 +1,35 @@
+/**
+ *  Copyright (C) 2017 3D Repo Ltd
+ *
+ *  This program is free software: you can redistribute it and/or modify
+ *  it under the terms of the GNU Affero General Public License as
+ *  published by the Free Software Foundation, either version 3 of the
+ *  License, or (at your option) any later version.
+ *
+ *  This program is distributed in the hope that it will be useful,
+ *  but WITHOUT ANY WARRANTY; without even the implied warranty of
+ *  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ *  GNU Affero General Public License for more details.
+ *
+ *  You should have received a copy of the GNU Affero General Public License
+ *  along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+
+import { PROJECT_ROLES_TYPES } from '../../constants/project-permissions';
+import { MODEL_ROLES_TYPES } from '../../constants/model-permissions';
+
+export const extendTeamspacesInfo = (teamspaces = []) => teamspaces.reduce((teamspacesWithAdminAccess, account) => {
+	const { isProjectAdmin, isModelAdmin } = account.projects.reduce((flags, { permissions, models }) => {
+		flags.isProjectAdmin = permissions.includes(PROJECT_ROLES_TYPES.ADMINISTRATOR);
+		flags.isModelAdmin = models.some((model) => model.permissions.includes(MODEL_ROLES_TYPES.ADMINISTRATOR));
+		return flags;
+	}, {});
+
+	if (account.isAdmin || isProjectAdmin || isModelAdmin) {
+		teamspacesWithAdminAccess.push({
+			...account,
+			isProjectAdmin
+		});
+	}
+	return teamspacesWithAdminAccess;
+}, []);
