+++ frontend/modules/teamspace/teamspace.selectors.ts
@@ -1,4 +1,22 @@
+/**
+ *  Copyright (C) 2017 3D Repo Ltd
+ *
+ *  This program is free software: you can redistribute it and/or modify
+ *  it under the terms of the GNU Affero General Public License as
+ *  published by the Free Software Foundation, either version 3 of the
+ *  License, or (at your option) any later version.
+ *
+ *  This program is distributed in the hope that it will be useful,
+ *  but WITHOUT ANY WARRANTY; without even the implied warranty of
+ *  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ *  GNU Affero General Public License for more details.
+ *
+ *  You should have received a copy of the GNU Affero General Public License
+ *  along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+
 import { createSelector } from 'reselect';
+import { extendTeamspacesInfo } from './teamspace.helpers';
 
 const selectTeamspaceDomain = (state) => Object.assign({}, state.teamspace);
 
@@ -10,8 +28,12 @@ export const selectCurrentUser = createSelector(
 	selectTeamspaceDomain, (state) => state.currentUser
 );
 
-export const selectCurrentUserTeamspaces = createSelector(
+export const selectTeamspaces = createSelector(
 	selectCurrentUser, (state) => state.accounts
+);
+
+export const selectTeamspacesWithAdminAccess = createSelector(
+	selectTeamspaces, (accounts) => extendTeamspacesInfo(accounts)
 );
 
 export const selectIsPending = createSelector(
